fix(fix-owner): throw away ENOENTs on chownr
This is a race npm has run into before. I am not sure how to
reproduce such a thing with cacache, but tbh if a directory
or file disappears mid-chown, I could care less?

Furthermore, if the file/directory's existence actually matters,
this will almost certainly error elsewhere. It's fine for us to
drop files on the floor though.