Use move-file instead of move-file-concurrently
Cacache only ever moves individual files, and move-file-concurrently
does a lot more than that, and has a lot of very outdated deps.

move-file does the same basic thing, but for a single file only, and
without the additional deps that are better to remove.

It does mean that, since copyFile is not an atomic operation, we can
end up files getting corrupted in some extremely unlikely situations,
but that will just be treated as a cache miss by most users anyhow.

Fix #33