fix: move to async functions where possible
This refactors the code to use async functions where possible.
It also uses `@npmcli/fs` consistently (it was already a dependency,
just not used).
It also reorders the test files to match their sources in `./lib`.

Tests were not refactored (except where needed to move to `@npmcli/fs`)
in the interest of an easier PR review.