fix: remove unused tar deendency
It appears that tar is declared in dependencies, but isn't actually used. This causes 2 MB of dependencies to be pulled into one's node_modules on every cacache install.

I was unable to track back when the dependency actually stopped being used; looks like at least several major versions were shipped with it unnecessarily.

If this is confirmed, it'd be great to push this fix to older major versions if possible.